requirements.txt
playwright==1.44.0
pytest==8.2.2
pytest-playwright
pytest-html
requests==2.32.3
python-dotenv==1.0.1
